Cache implicit receivers
